scrapers/EmbeddedMetadata/requirements.txt
pyexiv2
pyexiftool
stashapp-tools